move skill catalog builder into utils
